fix: move notifications into feature folder
