+++ src/ui/features/notifications/components/ErrorNotification.tsx
