+++ src/ui/features/notifications/components/InfoNotification.tsx
