+++ src/ui/features/notifications/components/WarningNotification.tsx
